Proto 1.2.2
Really Fixed Player Animations for Walking. Updated to stable and
experimental (Simply because stable was dated). Animations are now
contained in player.load(). Some animations aren't loading correctly
still, but that's a problem in either the file size of the sprite sheet
or the sprite code. Will work on creating Foliage. Ryan should be
updating a camera fix today or tommorow as he has time