Proto 1.2.3
Updated Player Abstraction significantly. Player input, equips, and
animations all calculated in modularized functions. This has the side
effect of cleaning player.draw and player.update. Will continue work in
the morning with the goal of finishing user input, and working on using
items on other objects in the world.